refactor(wire): extract contract policy leaf
Move wire contract and safety policy into a cycle-safe semantic leaf.
Keep compilation ordering, public call paths, and exclusion boundaries
unchanged.

All 69 definitions and 15 bindings are source- and AST-identical.

Validation: 859 tests passed on Pydantic 2.13.4 (95.87% coverage).
Validation: seven exclusion tests passed on Pydantic 2.12.3.
Validation: Ruff, ty, strict mypy, Vulture, and docs passed.
Validation: immutable v0.3.0 contract passed.

Closes #120
Refs #95
